requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
prometheus-client==0.20.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
rank-bm25==0.2.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
scikit-learn==1.4.2
numpy==1.26.4
Pillow==10.3.0
PyMuPDF==1.24.4
llama-cpp-python==0.2.79
